Ignoring contentFilter.  For some reason, this was coming in with values from another request, and causing the results to be blank.
For example, if a previous request had the 'path' set in contentFilter, it would set it globally.  That was overriding the
values from the collection query, and making it attempt to query a different path.  Thereby finding nothing.

Creating another value of `query` to the the same calculations.

Call this bug BUSTED!